requirements.txt
SimpleITK >= 2.1.1
pytorch >= 1.10.2